feat: Created health routes.
Used warp framework as it supports this use case while not sacrificing
performance. In particular actix has type issues that discourage
dependency injection and axum makes it difficult to manage state.

State is managed in with a service architecture that is injected into
each route handeler. This is a thread-safe operation using mutexes due
to the small scale of the data being stored and the infrequent updates
expected.